requirements.txt
google-generativeai
google-adk
qdrant-client
fastapi
uvicorn
python-dotenv
python-multipart
Pillow
transformers
torch
numpy
pydantic 